requirements.txt
ckanapi>=4.5
invoke>=1.5.0
pyyaml>=5.1
